refresh token subscribers

+++ src/HttpService.js
@@ -7,6 +7,9 @@ class HttpService {
         this.authStore = rootStore.authStore;
 
         this.clientID = 'L5EJWpimAOYOidk29pdoJyu2pdNjPkrHdpjeT2Vq';
+        this.refreshSubscribers = [];
+        this.isRefreshingToken = false;
+
         axios.defaults.baseURL = 'http://localhost:8003';
         axios.defaults.headers.common['Authorization'] = this.authStore.authToken;
 
@@ -24,15 +27,38 @@ class HttpService {
                     alert('로그인이 필요한 서비스입니다.');
                     this.rootStore.history.push('/login');
                 } else {
+                    if (!this.isRefreshingToken) {
+                        this.isRefreshingToken = true;
+                        return new Promise((resolve, reject) => {
+                            this.refreshToken().then(token => {
+                                originalRequest.headers.Authorization = this.authStore.authToken;
+                                resolve(axios(originalRequest));
+                                for (let subscriber of this.refreshSubscribers) {
+                                    subscriber(token);
+                                }
+                            }).catch(error => {
+                                this.authStore.deleteToken();
+                                reject(originalError);
+                                alert('로그인이 필요한 서비스입니다.');
+                                this.rootStore.history.push('/login');
+                                for (let subscriber of this.refreshSubscribers) {
+                                    subscriber(null);
+                                }
+                            }).finally(() => {
+                                this.refreshSubscribers = [];
+                                this.isRefreshingToken = false;
+                            });
+                        });
+                    }
+
                     return new Promise((resolve, reject) => {
-                        this.refreshToken().then(token => {
-                            originalRequest.headers.Authorization = this.authStore.authToken;
-                            resolve(axios(originalRequest));
-                        }).catch(error => {
-                            this.authStore.deleteToken();
-                            reject(originalError);
-                            alert('로그인이 필요한 서비스입니다.');
-                            this.rootStore.history.push('/login');
+                        this.refreshSubscribers.push(token => {
+                            if (token == null) {
+                                reject(originalError);
+                            } else {
+                                originalRequest.headers.Authorization = this.authStore.authToken;
+                                resolve(axios(originalRequest));
+                            }
                         });
                     });
                 }
